+++ src/euroeval/data_models.py
@@ -28,6 +28,8 @@ from .types import ScoreDict
 from .utils import get_package_version
 
 if t.TYPE_CHECKING:
+    from vllm.v1.attention.backends.registry import AttentionBackendEnum
+
     from .enums import InferenceBackend
 
 
@@ -517,6 +519,9 @@ class BenchmarkConfig:
             faster evaluation, but at the risk of running out of GPU memory. Only reduce
             this if you are running out of GPU memory. Only relevant if the model is
             generative.
+        attention_backend:
+            The attention backend to use for vLLM. Defaults to FLASHINFER. Only
+            relevant if the model is generative.
         requires_safetensors:
             Whether to only allow models that use the safetensors format.
         generative_type:
@@ -553,6 +558,7 @@ class BenchmarkConfig:
     few_shot: bool
     num_iterations: int
     gpu_memory_utilization: float
+    attention_backend: "AttentionBackendEnum"
     requires_safetensors: bool
     generative_type: GenerativeType | None
     download_only: bool
@@ -601,6 +607,7 @@ class BenchmarkConfigParams(pydantic.BaseModel):
     requires_safetensors: bool
     download_only: bool
     gpu_memory_utilization: float
+    attention_backend: str
     generative_type: GenerativeType | None
     custom_datasets_file: Path
     force: bool