+++ tests/conftest.py
@@ -83,6 +83,8 @@ def benchmark_config(
     auth: str, device: torch.device
 ) -> Generator[BenchmarkConfig, None, None]:
     """Yields a benchmark configuration used in tests."""
+    from vllm.v1.attention.backends.registry import AttentionBackendEnum
+
     yield BenchmarkConfig(
         languages=[DANISH],
         datasets=list(
@@ -107,6 +109,7 @@ def benchmark_config(
         api_base=None,
         api_version=None,
         gpu_memory_utilization=0.8,
+        attention_backend=AttentionBackendEnum.FLASHINFER,
         generative_type=None,
         debug=False,
         run_with_cli=True,